+++ beets/ui/_common.py
@@ -1,0 +1,4 @@
+class UserError(Exception):
+    """UI exception. Commands should throw this in order to display
+    nonrecoverable errors to the user.
+    """
